pyproject.toml
[tool.poetry]
name = "travel-planing-server"
version = "0.1.0"
description = ""
authors = ["shin-hama <htysivista@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
functions-framework = "^3.0.0"
Flask = "^2.1.2"
googlemaps = "^4.6.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.950"
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
